refactor: Padronizando melhor as pastas de extensao e constantes
