+++ src/OmegaFY.Chat.API.Domain/Constants/UserConstants.cs
@@ -1,4 +1,4 @@
-﻿namespace OmegaFY.Chat.API.Domain.Entities.Constants;
+﻿namespace OmegaFY.Chat.API.Domain.Constants;
 
 public class UserConstants
 {
