+++ src/OmegaFY.Chat.API.Infra/Authentication/Services/Implementations/IdentityAuthenticationService.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using OmegaFY.Chat.API.Common.Exceptions;
-using OmegaFY.Chat.API.Infra.Authentication.Extensions;
 using OmegaFY.Chat.API.Infra.Authentication.Models;
+using OmegaFY.Chat.API.Infra.Extensions;
 
 namespace OmegaFY.Chat.API.Infra.Authentication.Services.Implementations;
