+++ src/OmegaFY.Chat.API.Infra/Constants/OpenTelemetryConstants.cs
@@ -1,4 +1,4 @@
-﻿namespace OmegaFY.Chat.API.Infra.OpenTelemetry.Constants;
+﻿namespace OmegaFY.Chat.API.Infra.Constants;
 
 public static class OpenTelemetryConstants
 {
