+++ src/OmegaFY.Chat.API.WebAPI/DependencyInjection/Registrations/CacheRegistration.cs
@@ -1,4 +1,4 @@
-﻿using OmegaFY.Chat.API.Infra.DependencyInjection;
+﻿using OmegaFY.Chat.API.Infra.Extensions;
 
 namespace OmegaFY.Chat.API.WebAPI.DependencyInjection.Registrations;
