+++ src/OmegaFY.Chat.API.WebAPI/DependencyInjection/Registrations/HandlersRegistration.cs
@@ -1,6 +1,6 @@
 ﻿
 using OmegaFY.Chat.API.Application.DependencyInjection;
-using OmegaFY.Chat.API.Infra.DependencyInjection;
+using OmegaFY.Chat.API.Infra.Extensions;
 
 namespace OmegaFY.Chat.API.WebAPI.DependencyInjection.Registrations;
