+++ src/OmegaFY.Chat.API.WebAPI/DependencyInjection/Registrations/IdentityUserRegistration.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity;
-using OmegaFY.Chat.API.Data.EF.DependencyInjection;
-using OmegaFY.Chat.API.Infra.DependencyInjection;
+using OmegaFY.Chat.API.Data.EF.Extensions;
+using OmegaFY.Chat.API.Infra.Extensions;
 
 namespace OmegaFY.Chat.API.WebAPI.DependencyInjection.Registrations;
