+++ src/OmegaFY.Chat.API.WebAPI/Extensions/ActionExecutingContextExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc.Filters;
 using OmegaFY.Chat.API.Application.Shared;
 
-namespace OmegaFY.Chat.API.WebAPI.Filters.Extensions;
+namespace OmegaFY.Chat.API.WebAPI.Extensions;
 
 public static class ActionExecutingContextExtensions
 {
