+++ src/OmegaFY.Chat.API.WebAPI/OmegaFY.Chat.API.WebAPI.csproj
@@ -9,7 +9,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.8" />
-    <PackageReference Include="Scalar.AspNetCore" Version="2.6.9" />
+    <PackageReference Include="Scalar.AspNetCore" Version="2.7.2" />
   </ItemGroup>
 
   <ItemGroup>
@@ -17,10 +17,6 @@
     <ProjectReference Include="..\OmegaFY.Chat.API.Common\OmegaFY.Chat.API.Common.csproj" />
     <ProjectReference Include="..\OmegaFY.Chat.API.Data.EF\OmegaFY.Chat.API.Data.EF.csproj" />
     <ProjectReference Include="..\OmegaFY.Chat.API.Infra\OmegaFY.Chat.API.Infra.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Middlewares\" />
   </ItemGroup>
 
 </Project>
